FE: Move "model" directory up / remove "shared"
