+++ frontend/src/app/model/biography.ts
