+++ frontend/src/app/model/meme-preference.ts
