+++ frontend/src/app/model/meme.ts
