+++ frontend/src/app/model/message.ts
