+++ frontend/src/app/service/message.service.spec.ts
@@ -3,7 +3,7 @@ import { MessageService } from './message.service';
 import { HttpClientTestingModule,
   HttpTestingController } from '@angular/common/http/testing';
 import { UrlConfig } from '../url.config';
-import { UserMessage } from '../shared/model/message';
+import { UserMessage } from '../model/message';
 
 
 describe('MessageService', () => {
